+++ CHANGELOG.md
@@ -97,3 +97,4 @@
 - Enh #93: Introduce `Language` enum for standardized HTML lang codes and enhance `LangProvider` for testing (@terabytesoftw)
 - Enh #94: Add `HasDir` trait with `dir()` method and tests (@terabytesoftw)
 - Bug #95: Refactor exception handling in `HasDataTest` for improved clarity and consistency (@terabytesoftw)
+- Enh #96: Add `HasContentEditable` trait with `contentEditable()` method and tests (@terabytesoftw)